Rename task file
More appropriate now to not call it `_config.py`
