+++ src/cupbearer/tasks/__init__.py
@@ -1,4 +1,4 @@
 # ruff: noqa: F401
-from ._config import Task
 from .adversarial_examples import adversarial_examples
 from .backdoor_detection import backdoor_detection
+from .task import Task
