+++ src/cupbearer/tasks/backdoor_detection.py
@@ -3,7 +3,7 @@ from torch.utils.data import Dataset
 from cupbearer.data import Backdoor, BackdoorDataset
 from cupbearer.models import HookedModel
 
-from ._config import Task
+from .task import Task
 
 
 def backdoor_detection(
